scripts/requirements.txt
cryptography
PyJWT
requests
